requirements/base.txt
Django==2.0.4
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
bpython==0.17.1
celery[redis]==4.0.2
django-s3-storage==0.12.4
djangorestframework==3.8.2
django-rest-swagger==2.1.2
django-grappelli==2.11.1
